update some scripts

+++ docker/ipfs-docker-entrypoint.sh
@@ -8,21 +8,24 @@ if [[ "${no_conf}" != "1" ]]; then
   if [ ! -f /root/.ipfs ]; then 
     source /opt/paidpiper/ipfs.${PP_ENV}.cfg
     echo /onion3/${hsHostname}:4001/p2p/${ipfsSuperPeerID}
-    rm -rf /root/.ipfs
-    if [[ "${role}" == "hs_client" ]]; then
-      selfHsHostname="$(sed 's/[.].*$//' /root/tor/hidden_service/hsv3/hostname)"
-      /opt/paidpiper/ipfs init --announce=/onion3/${selfHsHostname}:4001 \
-          --bootStrap=/onion3/${hsHostname}:4001/p2p/${ipfsSuperPeerID} \
-          --torPath=/usr/local/bin/tor \
-          --torConfigPath=/usr/local/etc/tor/torrc
-      echo "Run as node"
-    else 
-      /opt/paidpiper/ipfs init \
-          --bootStrap=/onion3/${hsHostname}:4001/p2p/${ipfsSuperPeerID} \
-          --torPath=/usr/local/bin/tor \
-          --torConfigPath=/usr/local/etc/tor/torrc
-      echo "Run as client "
-    fi  
+    if [ ! -d "/root/.ipfs" ] 
+    then
+      rm -rf /root/.ipfs
+      if [[ "${role}" == "hs_client" ]]; then
+        selfHsHostname="$(sed 's/[.].*$//' /root/tor/hidden_service/hsv3/hostname)"
+        /opt/paidpiper/ipfs init --announce=/onion3/${selfHsHostname}:4001 \
+            --bootStrap=/onion3/${hsHostname}:4001/p2p/${ipfsSuperPeerID} \
+            --torPath=/usr/local/bin/tor \
+            --torConfigPath=/usr/local/etc/tor/torrc
+        echo "Run as node"
+      else 
+        /opt/paidpiper/ipfs init \
+            --bootStrap=/onion3/${hsHostname}:4001/p2p/${ipfsSuperPeerID} \
+            --torPath=/usr/local/bin/tor \
+            --torConfigPath=/usr/local/etc/tor/torrc
+        echo "Run as client "
+      fi  
+    fi
   fi  
 fi
 function mark {
